refactor error crate
